requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mutagen==1.47.0
pillow==11.1.0
psycopg2==2.9.10
PyNaCl==1.5.0
python-dotenv==1.0.1
solana==0.36.2
solders==0.23.0
sqlparse==0.5.3
typing_extensions==4.12.2
